requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
ansicon==1.89.0
async-timeout==4.0.2
attrs==21.4.0
aws-cdk-lib==2.46.0
aws-cdk.aws-lambda-python-alpha==2.46.0a0
black==22.10.0
blessed==1.19.1
boto3==1.24.66
botocore==1.27.66
cattrs==22.1.0
certifi==2022.6.15
chalice==1.27.3
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
constructs==10.1.94
contextlib2==21.6.0
defusedxml==0.7.1
distlib==0.3.6
exceptiongroup==1.0.0rc9
filelock==3.8.0
frozenlist==1.3.1
httpie==3.2.1
idna==3.3
inquirer==2.10.0
jinxed==1.2.0
jmespath==1.0.1
jsii==1.69.0
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
publication==0.0.3
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0
readchar==4.0.2
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
s3transfer==0.6.0
schema==0.7.5
six==1.16.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.4
wcwidth==0.2.5
yarl==1.8.1
